Make preliminary move to `batch tool`
From `tool list`
Deleted commands/tool & remove reference to it
Move code to commands/batch
