+++ src/adminware.py
@@ -5,7 +5,6 @@ from appliance_cli.commands import command_modules as standard_command_modules
 from commands import group
 from commands import batch
 from commands import open_command
-from commands import tool
 
 
 @click.group(help='Perform Flight Adminware management tasks.')
@@ -17,7 +16,6 @@ command_modules = standard_command_modules + [
     batch,
     group,
     open_command,
-    tool,
 ]
 
 for module in command_modules:
